Add workaround for bazel ObjC vs CcInfo provider issues
Currently when you link Swift binaries with swift_binary or swift_test
it uses the cc_common.link API. This API exclusively uses CcInfo
providers. When you use a rule from rules_apple, it links with bazel's
internal Objective-C specific logic, which ignores CcInfo providers.
Because of this cc_import doesn't add rpaths for the dependent library
when linking with Apple binaries. This change adds a custom rule that
propagates both the previous CcInfo that cc_import propagated (give or
take) along with the AppleFrameworkImportInfo and ObjC providers
necessary for Apple bundles to embed this dylib as a framework instead
of trying to load it as a dylib. The only case this doesn't currently
cover is the macOS CLI rule which doesn't support the cc_info path, but
also doesn't produce a bundle that can include the framework.